@@ -16,7 +16,8 @@ dependency_overrides:
     path: ../
 
 dev_dependencies:
-  e2e: ^0.2.1
+  integration_test:
+    path: ../../../integration_test
   flutter_driver:
     sdk: flutter
   pedantic: ^1.8.0
@@ -25,5 +26,5 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.4"
